+++ src/lib.rs
@@ -11,54 +11,32 @@ behavior.
 
 See [the documentation](https://docs.rs/big-brain) for more details.
 
-## Features
+### Features
 
 * Highly concurrent/parallelizable evaluation.
 * Integrates smoothly with Bevy.
-* Easy AI definition using idiomatic Rust builders. You don't have to be some genius to define behavior that _feels_ realistic to players.
-* High performance--supports hundreds of thousands of concurrent AIs.
-* Graceful degradation--can be configured such that the less frame time is available, the slower an AI might "seem", without dragging down framerates, by simply processing fewer events per tick.
 * Proven game AI model.
-* Low code overhead--you only define two types of application-dependent things, and everything else is building blocks!
 * Highly composable and reusable.
 * State machine-style continuous actions/behaviors.
 * Action cancellation.
 
-## Example
+### Example
 
-First, you define actions and considerations, which are just plain old Bevy
-Components and Systems. As a developer, you write application-dependent code
-to define [`Scorers`](#scorers) and [`Actions`](#actions), and then put it
-all together like building blocks, using [`Thinkers`](#thinkers) that will
-define the actual behavior.
+As a developer, you write application-dependent code to define
+[`Scorers`](#scorers) and [`Actions`](#actions), and then put it all together
+like building blocks, using [`Thinkers`](#thinkers) that will define the
+actual behavior.
 
-### Scorers
+#### Scorers
 
 `Scorer`s are entities that look at the world and evaluate into [`Score`](scorers::Score) values. You can think of them as the "eyes" of the AI system. They're a highly-parallel way of being able to look at the `World` and use it to make some decisions later.
 
-They are created by types that implement [`ScorerBuilder`](scorers::ScorerBuilder).
-
-```
+```rust
 use bevy::prelude::*;
 use big_brain::prelude::*;
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Component)]
 pub struct Thirsty;
-
-impl Thirsty {
-    fn build() -> ThirstyBuilder {
-        ThirstyBuilder
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ThirstyBuilder;
-
-impl ScorerBuilder for ThirstyBuilder {
-    fn build(&self, cmd: &mut Commands, scorer: Entity, _actor: Entity) {
-        cmd.entity(scorer).insert(Thirsty);
-    }
-}
 
 pub fn thirsty_scorer_system(
     thirsts: Query<&Thirst>,
@@ -72,31 +50,18 @@ pub fn thirsty_scorer_system(
 }
 ```
 
-### Actions
+#### Actions
 
-`Action`s are the actual things your entities will _do_. They are connected to [`ActionState`](actions::ActionState)s, and are created by types implementing [`ActionBuilder`](actions::ActionBuilder).
+`Action`s are the actual things your entities will _do_. They are connected to
+[`ActionState`](actions::ActionState)s that represent the current execution
+state of the state machine.
 
-```
+```rust
 use bevy::prelude::*;
 use big_brain::prelude::*;
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Component)]
 pub struct Drink;
-
-impl Drink {
-    pub fn build() -> DrinkBuilder {
-        DrinkBuilder
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct DrinkBuilder;
-
-impl ActionBuilder for DrinkBuilder {
-    fn build(&self, cmd: &mut Commands, action: Entity, _actor: Entity) {
-        cmd.entity(action).insert(Drink);
-    }
-}
 
 fn drink_action_system(
     mut thirsts: Query<&mut Thirst>,
@@ -119,32 +84,43 @@ fn drink_action_system(
 }
 ```
 
-### Thinkers
+#### Thinkers
 
 Finally, you can use it when define the [`Thinker`](thinker::Thinker), which you can attach as a
 regular Component:
 
-```no_run
+```rust
 cmd.spawn().insert(Thirst::new(70.0, 2.0)).insert(
     Thinker::build()
         .picker(FirstToScore { threshold: 0.8 })
-        .when(Thirsty::build(), Drink::build()),
+        .when(Thirsty, Drink),
 );
 ```
 
-## Contributing
+#### App
+
+Once all that's done, we just add our systems and off we go!
+
+```rust
+App::new()
+    .add_plugins(DefaultPlugins)
+    .add_plugin(BigBrainPlugin)
+    .add_startup_system(init_entities)
+    .add_system(thirst_system)
+    .add_system_to_stage(BigBrainStage::Actions, drink_action_system)
+    .add_system_to_stage(BigBrainStage::Scorers, thirsty_scorer_system)
+    .run();
+```
+
+### Contributing
 
 1. Install the latest Rust toolchain (stable supported).
 2. `cargo run --example thirst`
 3. Happy hacking!
 
-## License
+### License
 
-This project is licensed under [the Parity License](LICENSE.md). Third-party contributions are licensed under Apache-2.0 and belong to their respective authors.
-
-The Parity License is a copyleft license that, unlike the GPL family, allows you to license derivative and connected works under permissive licenses like MIT or Apache-2.0. It's free to use provided the work you do is freely available!
-
-For proprietary use, please [contact me](mailto:kzm@zkat.tech?subject=big-brain%20license), or just [sponsor me on GitHub](https://github.com/users/zkat/sponsorship) under the appropriate tier to [acquire a proprietary-use license](LICENSE-PATRON.md)! This funding model helps me make my work sustainable and compensates me for the work it took to write this crate!
+This project is licensed under [the Apache-2.0 License](LICENSE.md).
 
 */
 
@@ -152,7 +128,7 @@ pub mod evaluators;
 pub mod pickers;
 
 pub mod actions;
-mod choices;
+pub mod choices;
 pub mod scorers;
 pub mod thinker;
 
@@ -163,9 +139,12 @@ pub mod prelude {
     use super::*;
 
     pub use super::BigBrainPlugin;
-    pub use actions::{ActionBuilder, ActionState, Steps, Concurrently};
+    pub use super::BigBrainStage;
+    pub use actions::{ActionBuilder, ActionState, Concurrently, Steps};
     pub use pickers::{FirstToScore, Picker};
-    pub use scorers::{AllOrNothing, FixedScore, Score, ScorerBuilder, SumOfScorers, WinningScorer};
+    pub use scorers::{
+        AllOrNothing, FixedScore, Score, ScorerBuilder, SumOfScorers, WinningScorer,
+    };
     pub use thinker::{Actor, Thinker, ThinkerBuilder};
 }
 
@@ -186,20 +165,60 @@ App::build()
     // ...insert entities and other systems.
     .run();
 */
-
 pub struct BigBrainPlugin;
 
 impl Plugin for BigBrainPlugin {
-    fn build(&self, app: &mut AppBuilder) {
-        app.add_system(thinker::thinker_system.system());
-        app.add_system(thinker::thinker_component_attach_system.system());
-        app.add_system(thinker::thinker_component_detach_system.system());
-        app.add_system(thinker::actor_gone_cleanup.system());
-        app.add_system(actions::steps_system.system());
-        app.add_system(actions::concurrent_system.system());
-        app.add_system(scorers::fixed_score_system.system());
-        app.add_system(scorers::all_or_nothing_system.system());
-        app.add_system(scorers::sum_of_scorers_system.system());
-        app.add_system(scorers::winning_scorer_system.system());
+    fn build(&self, app: &mut App) {
+        use CoreStage::*;
+
+        app.add_stage_after(First, BigBrainStage::Scorers, SystemStage::parallel());
+        app.add_system_set_to_stage(
+            BigBrainStage::Scorers,
+            SystemSet::new()
+                .with_system(scorers::fixed_score_system)
+                .with_system(scorers::all_or_nothing_system)
+                .with_system(scorers::sum_of_scorers_system)
+                .with_system(scorers::winning_scorer_system)
+                .with_system(scorers::evaluating_scorer_system),
+        );
+
+        app.add_stage_after(
+            BigBrainStage::Scorers,
+            BigBrainStage::Thinkers,
+            SystemStage::parallel(),
+        );
+        app.add_system_to_stage(BigBrainStage::Thinkers, thinker::thinker_system);
+
+        app.add_stage_after(PreUpdate, BigBrainStage::Actions, SystemStage::parallel());
+        app.add_system_set_to_stage(
+            BigBrainStage::Actions,
+            SystemSet::new()
+                .with_system(actions::steps_system)
+                .with_system(actions::concurrent_system),
+        );
+
+        app.add_stage_after(Last, BigBrainStage::Cleanup, SystemStage::parallel());
+        app.add_system_set_to_stage(
+            BigBrainStage::Cleanup,
+            SystemSet::new()
+                .with_system(thinker::thinker_component_attach_system)
+                .with_system(thinker::thinker_component_detach_system)
+                .with_system(thinker::actor_gone_cleanup),
+        );
     }
+}
+
+/**
+BigBrainPlugin execution stages. Use these to schedule your own actions/scorers/etc.
+*/
+#[derive(Clone, Debug, Hash, Eq, PartialEq, StageLabel)]
+pub enum BigBrainStage {
+    /// Scorers are evaluated in this stage.
+    Scorers,
+    /// Actions are executed in this stage.
+    Actions,
+    /// Thinkers run their logic in this stage.
+    Thinkers,
+    /// Various internal cleanup items run in this final stage.
+    Cleanup,
 }