+++ src/scorers.rs
@@ -6,12 +6,15 @@ use std::{cmp::Ordering, sync::Arc};
 
 use bevy::prelude::*;
 
-use crate::thinker::{Actor, ScorerEnt};
+use crate::{
+    evaluators::Evaluator,
+    thinker::{Actor, ScorerEnt},
+};
 
 /**
 Score value between `0.0..=1.0` associated with a Scorer.
  */
-#[derive(Debug, Clone, Default)]
+#[derive(Clone, Component, Debug, Default)]
 pub struct Score(pub(crate) f32);
 
 impl Score {
@@ -43,12 +46,27 @@ The `build()` method MUST be implemented for any `ScorerBuilder`s you want to de
 */
 pub trait ScorerBuilder: std::fmt::Debug + Sync + Send {
     /**
-    MUST insert your concrete Scorer component into the Scorer [`Entity`], using `cmd`. You _may_ use `actor`, but it's perfectly normal to just ignore it.
+    MUST insert your concrete Scorer component into the Scorer [`Entity`], using
+     `cmd`. You _may_ use `actor`, but it's perfectly normal to just ignore it.
+
+    Note that this method is automatically implemented for any Components that
+    implement Clone, so you don't need to define it yourself unless you want
+    more complex parameterization of your Actions.
 
     ### Example
 
+    Using `Clone` (the easy way):
+
+    ```no_run
+    #[derive(Debug, Clone, Component)]
+    struct MyScorer;
+    ```
+
+    Implementing it manually:
+
     ```no_run
     struct MyBuilder;
+    #[derive(Debug, Component)]
     struct MyScorer;
 
     impl ScorerBuilder for MyBuilder {
@@ -60,10 +78,9 @@ pub trait ScorerBuilder: std::fmt::Debug + Sync + Send {
     */
     fn build(&self, cmd: &mut Commands, scorer: Entity, actor: Entity);
 
-    /**
-    Don't implement this yourself unless you know what you're doing.
-     */
-    fn attach(&self, cmd: &mut Commands, actor: Entity) -> Entity {
+    // Don't implement this yourself unless you know what you're doing.
+    #[doc(hidden)]
+    fn spawn_scorer(&self, cmd: &mut Commands, actor: Entity) -> Entity {
         let scorer_ent = cmd.spawn().id();
         cmd.entity(scorer_ent)
             .insert(Name::new("Scorer"))
@@ -74,30 +91,24 @@ pub trait ScorerBuilder: std::fmt::Debug + Sync + Send {
     }
 }
 
+impl<T> ScorerBuilder for T
+where
+    T: Component + Clone + std::fmt::Debug + Send + Sync,
+{
+    fn build(&self, cmd: &mut Commands, action: Entity, _actor: Entity) {
+        cmd.entity(action).insert(T::clone(self));
+    }
+}
+
 /**
 Scorer that always returns the same, fixed score. Good for combining with things creatively!
  */
-#[derive(Debug, Clone)]
+#[derive(Clone, Component, Debug)]
 pub struct FixedScore(f32);
-
-impl FixedScore {
-    pub fn build(score: f32) -> FixedScoreBuilder {
-        FixedScoreBuilder(score)
-    }
-}
 
 pub fn fixed_score_system(mut query: Query<(&FixedScore, &mut Score)>) {
     for (FixedScore(fixed), mut score) in query.iter_mut() {
         score.set(*fixed);
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct FixedScoreBuilder(f32);
-
-impl ScorerBuilder for FixedScoreBuilder {
-    fn build(&self, cmd: &mut Commands, action: Entity, _actor: Entity) {
-        cmd.entity(action).insert(FixedScore(self.0));
     }
 }
 
@@ -115,7 +126,7 @@ Thinker::build()
         MyAction::build());
 ```
  */
-#[derive(Debug)]
+#[derive(Component, Debug)]
 pub struct AllOrNothing {
     threshold: f32,
     scorers: Vec<ScorerEnt>,
@@ -174,10 +185,12 @@ impl ScorerBuilder for AllOrNothingBuilder {
         let scorers: Vec<_> = self
             .scorers
             .iter()
-            .map(|scorer| scorer.attach(cmd, actor))
+            .map(|scorer| scorer.spawn_scorer(cmd, actor))
             .collect();
         cmd.entity(scorer)
             .insert(Score::default())
+            .insert(Transform::default())
+            .insert(GlobalTransform::default())
             .push_children(&scorers[..])
             .insert(Name::new("Scorer"))
             .insert(AllOrNothing {
@@ -201,7 +214,7 @@ Thinker::build()
         MyAction::build());
 ```
  */
-#[derive(Debug)]
+#[derive(Component, Debug)]
 pub struct SumOfScorers {
     threshold: f32,
     scorers: Vec<ScorerEnt>,
@@ -252,15 +265,17 @@ impl SumOfScorersBuilder {
 }
 
 impl ScorerBuilder for SumOfScorersBuilder {
+    #[allow(clippy::needless_collect)]
     fn build(&self, cmd: &mut Commands, scorer: Entity, actor: Entity) {
         let scorers: Vec<_> = self
             .scorers
             .iter()
-            .map(|scorer| scorer.attach(cmd, actor))
+            .map(|scorer| scorer.spawn_scorer(cmd, actor))
             .collect();
         cmd.entity(scorer)
             .insert(Transform::default())
             .insert(GlobalTransform::default())
+            .push_children(&scorers[..])
             .insert(SumOfScorers {
                 threshold: self.threshold,
                 scorers: scorers.into_iter().map(ScorerEnt).collect(),
@@ -283,7 +298,7 @@ Thinker::build()
 ```
  */
 
-#[derive(Debug)]
+#[derive(Component, Debug)]
 pub struct WinningScorer {
     threshold: f32,
     scorers: Vec<ScorerEnt>,
@@ -300,13 +315,13 @@ impl WinningScorer {
 
 pub fn winning_scorer_system(
     mut query: Query<(Entity, &mut WinningScorer)>,
-    mut scores: QuerySet<(Query<&Score>, Query<&mut Score>)>,
+    mut scores: Query<&mut Score>,
 ) {
     for (sos_ent, mut winning_scorer) in query.iter_mut() {
         let (threshold, children) = (winning_scorer.threshold, &mut winning_scorer.scorers);
         let mut all_scores = children
             .iter()
-            .map(|ScorerEnt(e)| scores.q0().get(*e).expect("where is it?"))
+            .map(|ScorerEnt(e)| scores.get(*e).expect("where is it?"))
             .collect::<Vec<&Score>>();
 
         all_scores.sort_by(|a, b| a.get().partial_cmp(&b.get()).unwrap_or(Ordering::Equal));
@@ -320,7 +335,7 @@ pub fn winning_scorer_system(
             }
             None => 0.0,
         };
-        let mut score = scores.q1_mut().get_mut(sos_ent).expect("where did it go?");
+        let mut score = scores.get_mut(sos_ent).expect("where did it go?");
         score.set(crate::evaluators::clamp(winning_score_or_zero, 0.0, 1.0));
     }
 }
@@ -342,18 +357,92 @@ impl WinningScorerBuilder {
 }
 
 impl ScorerBuilder for WinningScorerBuilder {
+    #[allow(clippy::needless_collect)]
     fn build(&self, cmd: &mut Commands, scorer: Entity, actor: Entity) {
         let scorers: Vec<_> = self
             .scorers
             .iter()
-            .map(|scorer| scorer.attach(cmd, actor))
+            .map(|scorer| scorer.spawn_scorer(cmd, actor))
             .collect();
         cmd.entity(scorer)
             .insert(Transform::default())
             .insert(GlobalTransform::default())
+            .push_children(&scorers[..])
             .insert(WinningScorer {
                 threshold: self.threshold,
                 scorers: scorers.into_iter().map(ScorerEnt).collect(),
+            });
+    }
+}
+
+/**
+Composite scorer that takes a `ScorerBuilder` and applies an `Evaluator`. Note that
+unlike other composite scorers, `EvaluatingScorer` only takes one scorer upon building.
+
+### Example
+
+```ignore
+Thinker::build()
+    .when(
+        EvaluatingScorer::build(MyScorer, MyEvaluator),
+        MyAction);
+```
+ */
+#[derive(Clone, Component, Debug)]
+pub struct EvaluatingScorer {
+    scorer: ScorerEnt,
+    evaluator: Arc<dyn Evaluator>,
+}
+
+impl EvaluatingScorer {
+    pub fn build(
+        scorer: impl ScorerBuilder + 'static,
+        evaluator: impl Evaluator + 'static,
+    ) -> EvaluatingScorerBuilder {
+        EvaluatingScorerBuilder {
+            evaluator: Arc::new(evaluator),
+            scorer: Arc::new(scorer),
+        }
+    }
+}
+
+pub fn evaluating_scorer_system(
+    query: Query<(Entity, &EvaluatingScorer)>,
+    mut scores: Query<&mut Score>,
+) {
+    for (sos_ent, eval_scorer) in query.iter() {
+        // Get the inner score
+        let inner_score = scores
+            .get(eval_scorer.scorer.0)
+            .expect("where did it go?")
+            .get();
+        // Get composite score
+        let mut score = scores.get_mut(sos_ent).expect("where did it go?");
+        score.set(crate::evaluators::clamp(
+            eval_scorer.evaluator.evaluate(inner_score),
+            0.0,
+            1.0,
+        ));
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct EvaluatingScorerBuilder {
+    pub scorer: Arc<dyn ScorerBuilder>,
+    pub evaluator: Arc<dyn Evaluator>,
+}
+
+impl ScorerBuilder for EvaluatingScorerBuilder {
+    fn build(&self, cmd: &mut Commands, scorer: Entity, actor: Entity) {
+        let inner_scorer = self.scorer.spawn_scorer(cmd, actor);
+        let scorers = vec![inner_scorer];
+        cmd.entity(scorer)
+            .insert(Transform::default())
+            .insert(GlobalTransform::default())
+            .push_children(&scorers[..])
+            .insert(EvaluatingScorer {
+                evaluator: self.evaluator.clone(),
+                scorer: ScorerEnt(inner_scorer),
             });
     }
 }